feat: make login functional again
